Refactored package structure
