+++ src/main/java/org/wahlzeit_revisited/agent/NotifyAboutPraiseAgent.java
@@ -23,7 +23,7 @@ package org.wahlzeit_revisited.agent;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.NotFoundException;
 import org.wahlzeit_revisited.model.*;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.api.repository.UserRepository;
 import org.wahlzeit_revisited.service.mailing.EmailService;
 import org.wahlzeit_revisited.service.mailing.EmailServiceManager;
 import org.wahlzeit_revisited.utils.SysLog;
