+++ src/main/java/org/wahlzeit_revisited/api/filter/AuthenticationFilter.java
@@ -19,7 +19,7 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-package org.wahlzeit_revisited.filter;
+package org.wahlzeit_revisited.api.filter;
 
 import jakarta.annotation.Priority;
 import jakarta.annotation.security.DenyAll;
@@ -34,11 +34,11 @@ import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.SecurityContext;
 import jakarta.ws.rs.ext.Provider;
-import org.wahlzeit_revisited.dto.ErrorDto;
-import org.wahlzeit_revisited.auth.AccessRights;
-import org.wahlzeit_revisited.auth.PrincipalUser;
+import org.wahlzeit_revisited.api.dto.ErrorDto;
+import org.wahlzeit_revisited.api.auth.AccessRights;
+import org.wahlzeit_revisited.api.auth.PrincipalUser;
 import org.wahlzeit_revisited.model.User;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.api.repository.UserRepository;
 import org.wahlzeit_revisited.utils.SysLog;
 
 import java.lang.reflect.Method;
