+++ src/main/java/org/wahlzeit_revisited/api/repository/PersistentFactory.java
@@ -19,27 +19,19 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-package org.wahlzeit_revisited.repository;
+package org.wahlzeit_revisited.api.repository;
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.List;
-import java.util.Optional;
 
-/**
- * A Repository is a wrapper between the Database and Java POJOs
- * It provides CRUD operations on the entities
- * @param <T> to represent Persistent
- */
-public interface Repository<T extends Persistent> {
+public interface PersistentFactory<T extends Persistent> {
 
-    Optional<T> findById(Long id) throws SQLException;
-
-    List<T> findAll() throws SQLException;
-
-    T insert(T toInsert) throws SQLException;
-
-    T update(T toUpdate) throws SQLException;
-
-    T delete(T toDelete) throws SQLException;
+    /**
+     * Creates an persistent entity from a db query
+     * @param resultSet database row
+     * @return initialized entity
+     * @throws SQLException invalid SQL statement
+     */
+    T createPersistent(ResultSet resultSet) throws SQLException;
 
 }
