--- src/main/java/org/wahlzeit_revisited/dto/UserDto.java
@@ -1,90 +1,0 @@
-/*
- * Copyright (c) 2006-2009 by Dirk Riehle, http://dirkriehle.com
- * Copyright (c) 2021 by Aron Metzig
- *
- * This file is part of the Wahlzeit photo rating application.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public
- * License along with this program. If not, see
- * <http://www.gnu.org/licenses/>.
- */
-
-package org.wahlzeit_revisited.dto;
-
-/**
- * UserDto represents the User-entity for the outer world
- */
-public class UserDto {
-
-    private final Long id;
-    private final String name;
-    private final String email;
-
-    UserDto(Long id, String name, String email) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-    }
-
-    /**
-     * @methodtype get
-     */
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * @methodtype get
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * @methodtype get
-     */
-    public String getEmail() {
-        return email;
-    }
-
-    /*
-     * Builder
-     */
-
-    public static class Builder {
-        private Long id;
-        private String name;
-        private String email;
-
-        public Builder withId(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder withName(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public Builder withEmail(String email) {
-            this.email = email;
-            return this;
-        }
-
-        public UserDto build() {
-            return new UserDto(id, name, email);
-        }
-
-    }
-
-}
