+++ src/main/java/org/wahlzeit_revisited/model/Case.java
@@ -21,7 +21,7 @@
 package org.wahlzeit_revisited.model;
 
 
-import org.wahlzeit_revisited.repository.Persistent;
+import org.wahlzeit_revisited.api.repository.Persistent;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
