+++ src/main/java/org/wahlzeit_revisited/model/CaseFactory.java
@@ -21,7 +21,7 @@
 
 package org.wahlzeit_revisited.model;
 
-import org.wahlzeit_revisited.repository.PersistentFactory;
+import org.wahlzeit_revisited.api.repository.PersistentFactory;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
