+++ src/main/java/org/wahlzeit_revisited/model/Client.java
@@ -21,7 +21,7 @@
 package org.wahlzeit_revisited.model;
 
 
-import org.wahlzeit_revisited.auth.AccessRights;
+import org.wahlzeit_revisited.api.auth.AccessRights;
 
 /**
  * A Client uses the system. It is an abstract superclass. This package defines guest, user, moderator, and
