+++ src/main/java/org/wahlzeit_revisited/service/PhotoService.java
@@ -31,10 +31,10 @@ import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
 import org.wahlzeit_revisited.agent.AgentManager;
 import org.wahlzeit_revisited.agent.NotifyAboutPraiseAgent;
-import org.wahlzeit_revisited.dto.PhotoDto;
+import org.wahlzeit_revisited.api.dto.PhotoDto;
 import org.wahlzeit_revisited.model.*;
-import org.wahlzeit_revisited.repository.PhotoRepository;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.api.repository.PhotoRepository;
+import org.wahlzeit_revisited.api.repository.UserRepository;
 import org.wahlzeit_revisited.utils.SysLog;
 import org.wahlzeit_revisited.utils.WahlzeitConfig;
