+++ src/main/java/org/wahlzeit_revisited/service/UserService.java
@@ -26,11 +26,11 @@ import jakarta.inject.Singleton;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
-import org.wahlzeit_revisited.auth.AccessRights;
-import org.wahlzeit_revisited.dto.UserDto;
+import org.wahlzeit_revisited.api.auth.AccessRights;
+import org.wahlzeit_revisited.api.dto.UserDto;
 import org.wahlzeit_revisited.model.User;
 import org.wahlzeit_revisited.model.UserFactory;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.api.repository.UserRepository;
 import org.wahlzeit_revisited.utils.SysLog;
 
 import java.sql.SQLException;
