+++ src/test/java/org/wahlzeit_revisited/service/UserServiceIT.java
@@ -5,10 +5,10 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.wahlzeit_revisited.BaseModelTest;
-import org.wahlzeit_revisited.dto.UserDto;
+import org.wahlzeit_revisited.api.dto.UserDto;
 import org.wahlzeit_revisited.model.User;
 import org.wahlzeit_revisited.model.UserFactory;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.api.repository.UserRepository;
 
 import java.sql.SQLException;
 import java.util.List;
